Add the clock-data component.
The clock-data component allows for drift-free notifications of the
passing of time.

It requires :
 - an interval at which a tick will be triggered ;
 - an offset to the interval.

These parameters are durations to be parsed by the moment.js
moment.Duration constructor.

At each tick, the "tick" event will be fired ; it shall contain two
fields, each containing an ISO8601 date string :
{
    time: String,
    next: String
}

Also at each tick the "current" and "next" properties will be updated to
contain the ISO8601 date strings of the current and next ticks.