Add tags to the <room-planning/>.
<room-planning/> now has two additional properties : backup-rooms and
backup-tags.

backup-rooms is an array of room names. These rooms will be searched for
tagged events. If an event is tagged with a tag in backup-tags, an
additional column will be displayed for this room.

backup-tags is an array of strings. Each string represents a tag which
will trigger the display of additional rooms.